Added lock file controls to stop calls that interact with the aws cluster if another process is already interacting with it.

+++ stochss/handlers/util/scripts/aws_compute.py
@@ -17,9 +17,12 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 '''
+import os
 import sys
 import logging
 import argparse
+
+import fcntl
 
 sys.path.append("/stochss/stochss/") # pylint: disable=wrong-import-position
 sys.path.append("/stochss/stochss/handlers/") # pylint: disable=wrong-import-position
@@ -54,29 +57,49 @@ def launch_instance():
     Launch an AWS instance.
     '''
     base = StochSSBase(path="")
-    base.launch_aws_cluster()
+    if not is_file_locked(base.user_dir):
+        base.launch_aws_cluster()
+
+def is_file_locked(user_dir):
+    '''
+    Check if the aws lock file is already locked, if not lock it.
+
+    Attributes
+    ----------
+    user_dir : str
+        The users home directory.
+    '''
+    l_path = os.path.join(user_dir, '.aws/awsec2.lock')
+    with open(l_path, 'w', encoding="utf-8") as lock_fd:
+        try:
+            fcntl.lockf(lock_fd, fcntl.LOCK_EX | fcntl.LOCK_NB)
+            return False
+        except OSError:
+            return True
 
 def update_status():
     '''
     Update the status of an AWS instance.
     '''
     base = StochSSBase(path="")
-    settings = base.load_user_settings(path='.user-settings.json')
+    if not is_file_locked(base.user_dir):
+        settings = base.load_user_settings(path='.user-settings.json')
 
-    instance = settings['headNode']
-    s_path = f".aws/{instance.replace('.', '-')}-status.txt"
+        instance = settings['headNode']
+        s_path = f".aws/{instance.replace('.', '-')}-status.txt"
 
-    cluster = base.get_aws_cluster()
-    if cluster._server is not None:
-        with open(s_path, "w", encoding="utf-8") as aws_s_fd:
-            aws_s_fd.write(cluster._server.state['Name'])
+        cluster = base.get_aws_cluster()
+        if cluster._server is not None:
+            with open(s_path, "w", encoding="utf-8") as aws_s_fd:
+                aws_s_fd.write(cluster._server.state['Name'])
 
 def terminate_instance():
     '''
     Terminate an AWS instance.
     '''
     base = StochSSBase(path="")
-    base.terminate_aws_cluster()
+    if not is_file_locked(base.user_dir):
+        base.terminate_aws_cluster()
 
 if __name__ == "__main__":
     args = get_parsed_args()
